add context to pytest
