+++ src/gui/src/stores/AssessStore.js
@@ -203,7 +203,10 @@ export const useAssessStore = defineStore('assess', {
   persist: {
     paths: ['osint_sources', 'osint_source_groups', 'newsItems'],
     serializer: {
-      deserialize: (value) => JSON.parse(value),
+      deserialize: (value) => {
+        console.debug('Deserializing', value)
+        return JSON.parse(value)
+      },
       serialize: (value) => JSON.stringify(value)
     }
   }
